# docs/swagger.yaml
basePath: /
definitions:
...
  linkpulse_internal_models.ErrorDetails:
    properties:
      code:
        example: INVALID_CREDENTIALS
        type: string
      message:
        example: Invalid email or password
        type: string
    type: object
  linkpulse_internal_models.ErrorResponse:
    properties:
      detail:
        example: The email or password provided is incorrect.
        type: string
      error:
        $ref: '#/definitions/linkpulse_internal_models.ErrorDetails'
      instance:
        example: /api/v1/auth/login
        type: string
      request_id:
        example: uuid-request-id
        type: string
      status:
        example: 401
        type: integer
      success:
        example: false
        type: boolean
      title:
        example: Invalid Credentials
        type: string
      type:
        example: https://linkpulse.com/errors/invalid-credentials
        type: string
    type: object
...
  linkpulse_internal_models.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      request_id:
        example: uuid-request-id
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  linkpulse_internal_models.ValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: The email field is required
        type: string
      rule:
        example: required
        type: string
    type: object
  linkpulse_internal_models.ValidationErrorResponse:
    properties:
      detail:
        example: Validation failed
        type: string
      details:
        items:
          $ref: '#/definitions/linkpulse_internal_models.ValidationError'
        type: array
      error:
        $ref: '#/definitions/linkpulse_internal_models.ErrorDetails'
      instance:
        example: /api/v1/links
        type: string
      request_id:
        example: uuid-request-id
        type: string
      status:
        example: 422
        type: integer
      success:
        example: false
        type: boolean
      title:
        example: Validation Error
        type: string
      type:
        example: https://linkpulse.com/errors/validation-error
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: JyotiBhardwajj
    url: https://github.com/JyotiBhardwajj/LinkPulse
  description: |-
    Production-grade URL shortener with analytics, authentication, and observability.

    ## Authentication
    All protected endpoints require a Bearer JWT in the Authorization header.
    Obtain tokens via POST /api/v1/auth/login.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LinkPulse API
  version: "1.0"
paths:
  /api/v1/analytics/browsers:
    get:
      description: Returns click counts grouped by browser (Chrome, Firefox, Safari,
        etc.).
      parameters:
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Browser distribution
      tags:
      - analytics
  /api/v1/analytics/clicks:
    get:
      description: Returns time-series click data grouped by interval for the authenticated
        user's links.
      parameters:
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      - description: 'Grouping interval: hour, day, week'
        in: query
        name: interval
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ValidationErrorResponse'
      security:
      - BearerAuth: []
      summary: Clicks over time
      tags:
      - analytics
  /api/v1/analytics/devices:
    get:
      description: Returns click counts grouped by device type (mobile, desktop, tablet,
        etc.).
      parameters:
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Device distribution
      tags:
      - analytics
  /api/v1/analytics/overview:
    get:
      description: Returns aggregate statistics for all links owned by the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Analytics overview
      tags:
      - analytics
  /api/v1/analytics/referrers:
    get:
      description: Returns click counts grouped by referrer domain.
      parameters:
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      - description: Max referrers to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Referrer distribution
      tags:
      - analytics
  /api/v1/analytics/top-links:
    get:
      description: Returns the top N links ranked by click count for the authenticated
        user.
      parameters:
      - description: Number of results (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top performing links
      tags:
      - analytics
...
  /api/v1/links/{id}/analytics:
    get:
      description: Returns detailed time-series analytics for a specific link owned
        by the authenticated user.
      parameters:
      - description: Link UUID
        in: path
        name: id
        required: true
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Link analytics
      tags:
      - links
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"